Add Java / Kotlin bindings for SurfaceOrientation.
Note that libgeometry is already a part of `filament-android`, this
simply exposes its functionality in a better way.

The approach we're taking with Java-land SurfaceOrientation is similar
to the its JavaScript counterpart in that we have bundled its language
bindings into the main Filament package, even though it is a separate
library in C++ land. This is much simpler than creating a brand new Java
package.

Fixes #1729.